Added transfer code generator with checksum
First step towards https://phabricator.wikimedia.org/T170266

Has checkstum and 8 characeters instead of 7 to add some extra entropy.
Like this we have 150 billion combinations, which should give us enough
room for adding a bad word filter.

Added a dash in the center of the thing for readability.

Checksum uses md5. This is the simplest thing I could think off.
One drawback is that it can include numbers, unlike the rest of the
code. Not sure if that is a thing we want to avoid or not.

Did not include the prefix in this class as I think this does not
belong here if it depends on something else... can add it back
later if this turns out to make sense when we get to that point.